backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
pillow==11.0.0
babel==2.16.0
pytz==2024.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
python-dotenv==1.0.1
django-storages==1.14.4
boto3==1.35.22





